requirements.txt
Flask==2.3.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
transformers==4.40.0
torch==2.3.0
reportlab==4.1.0
gunicorn==21.2.0